Disable otel extensions for dev images
They never should have enabled in those
