+++ Dockerfile-zts-alpine
@@ -77,7 +77,7 @@ RUN EXTENSION_DIR=`php-config --extension-dir 2>/dev/null` && \
         $PHPIZE_DEPS \
     ## Install PECL
     && wget -q pear.php.net/go-pear.phar && php go-pear.phar \
-    && install-php-extensions pcntl pgsql pdo pdo_pgsql bcmath zip gmp iconv opcache intl sockets ffi opentelemetry grpc \
+    && install-php-extensions pcntl pgsql pdo pdo_pgsql bcmath zip gmp iconv opcache intl sockets \
     && (install-php-extensions random || true) \
     && pecl install parallel || pecl install parallel-1.1.4 \
     && docker-php-ext-enable parallel \
@@ -144,7 +144,9 @@ USER app
 
 ## ZTS stage ##
 FROM zts-slim-root AS zts-slim
+RUN install-php-extensions ffi opentelemetry grpc
 USER app
 
 FROM zts-root AS zts
+RUN install-php-extensions ffi opentelemetry grpc
 USER app
