+++ Dockerfile-zts-debian
@@ -72,7 +72,7 @@ RUN EXTENSION_DIR=`php-config --extension-dir 2>/dev/null` && \
         wget \
         gdb \
         $PHPIZE_DEPS \
-    && docker-php-ext-install -j$(nproc) pcntl pgsql pdo pdo_pgsql bcmath zip gmp iconv intl sockets ffi opentelemetry grpc \
+    && docker-php-ext-install -j$(nproc) pcntl pgsql pdo pdo_pgsql bcmath zip gmp iconv intl sockets \
     && (docker-php-ext-install -j$(nproc) random || true) \
     && pecl install parallel || pecl install parallel-1.1.4 \
     && docker-php-ext-enable parallel \
@@ -148,7 +148,9 @@ USER app
 
 ## ZTS stage ##
 FROM zts-slim-root AS zts-slim
+RUN docker-php-ext-install -j$(nproc) ffi opentelemetry grpc
 USER app
 
 FROM zts-root AS zts
+RUN docker-php-ext-install -j$(nproc) ffi opentelemetry grpc
 USER app
